trinethra/services/explain_service/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
psycopg[binary]==3.3.3